+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 ### Fixed
 
 - Race condition in `am_ao_publish_exclude_x()`
+- Fix global start-up gate to avoid startup race conditions in preemptive builds
 
 ## v0.15.1 - 28-March-2026